Applied changes from #529, fixing #508 - bugs in migration packages.
Moved the failing test to main gnrsqlmigration test module, and removed the
fail status, in a generic code test.

Updated runtime imports.

close #508
